Fixed up/renamed plugin groovy code
