+++ plugins/nf-iridanext/src/main/nextflow/iridanext/IridaNextFactory.groovy
@@ -1,5 +1,6 @@
 /*
- * Copyright 2021, Seqera Labs
+ * Original file Copyright 2021, Seqera Labs (from nf-hello plugin template)
+ * Modifications Copyright 2023, Government of Canada
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,22 +15,25 @@
  * limitations under the License.
  */
 
-package nextflow.hello
+package nextflow.iridanext
 
 import groovy.transform.CompileStatic
-import nextflow.plugin.BasePlugin
-import nextflow.plugin.Scoped
-import org.pf4j.PluginWrapper
-
+import nextflow.Session
+import nextflow.trace.TraceObserver
+import nextflow.trace.TraceObserverFactory
 /**
- * Implements the Hello plugins entry point
+ * Factory for building the IridaNextObserver
  *
+ * @author Aaron Petkau <aaron.petkau@phac-aspc.gc.ca>
  * @author Paolo Di Tommaso <paolo.ditommaso@gmail.com>
  */
 @CompileStatic
-class HelloPlugin extends BasePlugin {
+class IridaNextFactory implements TraceObserverFactory {
 
-    HelloPlugin(PluginWrapper wrapper) {
-        super(wrapper)
+    @Override
+    Collection<TraceObserver> create(Session session) {
+        final result = new ArrayList()
+        result.add( new IridaNextObserver() )
+        return result
     }
 }
